# docs/python-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StaffMatch Python API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/matching/calculate": {
      "post": {
        "summary": "Calculate Matching",
        "operationId": "calculate_matching_api_v1_matching_calculate_post",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Request-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchingResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/tests/generate": {
      "post": {
        "summary": "Generate Test",
        "operationId": "generate_test_api_v1_tests_generate_post",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Request-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/tests/score": {
      "post": {
        "summary": "Score Test",
        "operationId": "score_test_api_v1_tests_score_post",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Request-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/api/v1/ranking/calculate": {
      "post": {
        "summary": "Calculate Ranking",
        "operationId": "calculate_ranking_api_v1_ranking_calculate_post",
        "parameters": [
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Request-Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RankingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Candidate": {
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Candidateid"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "workFormats": {
            "items": {
              "type": "string",
              "enum": [
                "REMOTE",
                "HYBRID",
                "OFFICE"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Workformats"
          },
          "availableHours": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 168.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Availablehours"
          },
          "salaryExpectation": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salaryexpectation"
          },
          "experienceMonths": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiencemonths"
          },
          "portfolioAvailable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolioavailable"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/CandidateSkill"
            },
            "type": "array",
            "minItems": 1,
            "title": "Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidateId",
          "workFormats",
          "skills"
        ],
        "title": "Candidate"
      },
...
      "ChoiceQuestion": {
        "properties": {
          "questionId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Questionid"
          },
          "type": {
            "type": "string",
            "const": "SINGLE_CHOICE",
            "title": "Type",
            "default": "SINGLE_CHOICE"
          },
          "competency": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Competency"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Text"
          },
          "options": {
            "items": {
              "type": "string",
              "minLength": 1,
              "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]"
            },
            "type": "array",
            "minItems": 2,
            "title": "Options"
          },
          "correctAnswer": {
            "$ref": "#/components/schemas/CorrectAnswer"
          },
          "maxScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Maxscore"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "questionId",
          "competency",
          "text",
          "options",
          "correctAnswer",
          "maxScore"
        ],
        "title": "ChoiceQuestion"
      },
      "Competency": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Code"
          },
          "weight": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Weight"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "weight"
        ],
        "title": "Competency"
      },
...
      "ErrorDetail": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "requestId": {
            "type": "string",
            "title": "Requestid"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "requestId",
          "status",
          "error",
          "message",
          "retryable",
          "details"
        ],
        "title": "ErrorResponse"
      },
...
      "GenerateRequest": {
        "properties": {
          "vacancyId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Vacancyid"
          },
          "mode": {
            "type": "string",
            "const": "AUTO",
            "title": "Mode"
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Role"
          },
          "level": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Level"
          },
          "durationMinutes": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 5.0,
            "title": "Durationminutes"
          },
          "questionCount": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 6.0,
            "title": "Questioncount"
          },
          "competencies": {
            "items": {
              "$ref": "#/components/schemas/Competency"
            },
            "type": "array",
            "minItems": 1,
            "title": "Competencies"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vacancyId",
          "mode",
          "role",
          "level",
          "durationMinutes",
          "questionCount",
          "competencies"
        ],
        "title": "GenerateRequest"
      },
      "GenerateResponse": {
        "properties": {
          "testId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Testid"
          },
          "generationVersion": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Generationversion"
          },
          "estimatedDurationMinutes": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 5.0,
            "title": "Estimateddurationminutes"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ChoiceQuestion"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "testId",
          "generationVersion",
          "estimatedDurationMinutes",
          "questions"
        ],
        "title": "GenerateResponse"
      },
      "Match": {
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Candidateid"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "matched": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "type": "array",
            "title": "Matched"
          },
          "partial": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "type": "array",
            "title": "Partial"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/Explanation"
            },
            "type": "array",
            "title": "Missing"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidateId",
          "eligible",
          "score",
          "matched",
          "partial",
          "missing"
        ],
        "title": "Match"
      },
      "MatchingRequest": {
        "properties": {
          "vacancy": {
            "$ref": "#/components/schemas/Vacancy"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Candidates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vacancy",
          "candidates"
        ],
        "title": "MatchingRequest"
      },
      "MatchingResponse": {
        "properties": {
          "vacancyId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Vacancyid"
          },
          "algorithmVersion": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Algorithmversion"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/Match"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vacancyId",
          "algorithmVersion",
          "matches"
        ],
        "title": "MatchingResponse"
      },
...
      "RankingRequest": {
        "properties": {
          "vacancyId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Vacancyid"
          },
          "weights": {
            "$ref": "#/components/schemas/Weights"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RankCandidate"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vacancyId",
          "candidates"
        ],
        "title": "RankingRequest"
      },
      "RankingResponse": {
        "properties": {
          "vacancyId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Vacancyid"
          },
          "rankingVersion": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Rankingversion"
          },
          "final": {
            "items": {
              "$ref": "#/components/schemas/FinalEntry"
            },
            "type": "array",
            "title": "Final"
          },
          "waiting": {
            "items": {
              "$ref": "#/components/schemas/WaitingEntry"
            },
            "type": "array",
            "title": "Waiting"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "vacancyId",
          "rankingVersion",
          "final",
          "waiting"
        ],
        "title": "RankingResponse"
      },
...
      "ScoreQuestion": {
        "properties": {
          "questionId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Questionid"
          },
          "type": {
            "type": "string",
            "enum": [
              "SINGLE_CHOICE",
              "FREE_TEXT"
            ],
            "title": "Type"
          },
          "competency": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Competency"
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "correctAnswer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CorrectAnswer"
              },
              {
                "type": "null"
              }
            ]
          },
          "rubric": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rubric"
              },
              {
                "type": "null"
              }
            ]
          },
          "maxScore": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Maxscore"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "questionId",
          "type",
          "competency",
          "maxScore"
        ],
        "title": "ScoreQuestion"
      },
      "ScoreRequest": {
        "properties": {
          "assignmentId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Assignmentid"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/ScoreQuestion"
            },
            "type": "array",
            "minItems": 1,
            "title": "Questions"
          },
          "answers": {
            "items": {
              "$ref": "#/components/schemas/Answer"
            },
            "type": "array",
            "title": "Answers"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assignmentId",
          "questions",
          "answers"
        ],
        "title": "ScoreRequest"
      },
      "ScoreResponse": {
        "properties": {
          "assignmentId": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Assignmentid"
          },
          "scoringVersion": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Scoringversion"
          },
          "modelVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelversion"
          },
          "totalScore": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Totalscore"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResult"
            },
            "type": "array",
            "title": "Questions"
          },
          "breakdown": {
            "items": {
              "$ref": "#/components/schemas/Breakdown"
            },
            "type": "array",
            "title": "Breakdown"
          },
          "summary": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "assignmentId",
          "scoringVersion",
          "totalScore",
          "questions",
          "breakdown",
          "summary"
        ],
        "title": "ScoreResponse"
      },
      "Vacancy": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "pattern": "[^\\t-\\r\\u001c-\\u0020\\u1680\\u2000-\\u2006\\u2008-\\u200a\\u2028\\u2029\\u205f\\u3000]",
            "title": "Role"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "workFormat": {
            "type": "string",
            "enum": [
              "REMOTE",
              "HYBRID",
              "OFFICE"
            ],
            "title": "Workformat"
          },
          "hoursMin": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 168.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hoursmin"
          },
          "salaryFrom": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salaryfrom"
          },
          "salaryTo": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salaryto"
          },
          "experienceMonthsMin": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Experiencemonthsmin"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/VacancySkill"
            },
            "type": "array",
            "minItems": 1,
            "title": "Skills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "role",
          "workFormat",
          "skills"
        ],
        "title": "Vacancy"
      },
...
    }
  }
}